+++ raiden/network/transport/matrix/rtc/aio_queue.py
@@ -1,12 +1,15 @@
 from asyncio import Future
-from typing import Any, Callable, Dict
+from dataclasses import dataclass
+from typing import Any, Callable, Dict, Optional
 
 import gevent
+from aiortc import RTCDataChannel, RTCPeerConnection
 from gevent.lock import Semaphore
 from gevent.queue import Empty, Queue
 
 from raiden.network.transport.matrix.rtc import aiogevent
-from raiden.network.transport.matrix.rtc.web_rtc import RTCPartner
+from raiden.network.transport.matrix.utils import my_place_or_yours
+from raiden.utils.formatting import to_checksum_address
 from raiden.utils.typing import Address
 
 
@@ -15,6 +18,21 @@ def make_wrapped_greenlet(target: Callable, *args: Any, **kwargs: Any) -> Future
     wrapped_glet = aiogevent.wrap_greenlet(glet)
     glet.start()
     return wrapped_glet
+
+
+@dataclass
+class RTCPartner:
+    partner_address: Address
+    pc: RTCPeerConnection
+    channel: Optional[RTCDataChannel] = None
+
+    def create_channel(self, node_address) -> None:
+        lower_address = my_place_or_yours(node_address, self.partner_address)
+        higher_address = self.partner_address if lower_address == node_address else node_address
+        channel_name = (
+            f"{to_checksum_address(lower_address)}|{to_checksum_address(higher_address)}"
+        )
+        self.channel = self.pc.createDataChannel(channel_name)
 
 
 class AGTransceiver:
